backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.1
anthropic==0.28.0
Pillow==10.3.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
numpy==1.26.4
redis==5.0.4
celery==5.4.0
boto3==1.34.114
aiofiles==23.2.1
python-dotenv==1.0.1
sentry-sdk[fastapi]==2.3.1
prometheus-fastapi-instrumentator==7.0.0
email-validator==2.1.1
google-generativeai>=0.8.0
